requirements.txt
AccessControl==5.7
Acquisition==4.13
AuthEncoding==4.3
beautifulsoup4==4.11.1
BTrees==4.11.3
cffi==1.15.1
Chameleon==3.10.1
DateTime==4.7
DBUtils==3.0.2
DocumentTemplate==4.1
ExtensionClass==4.9
grpcio==1.51.1
grpcio-tools==1.51.1
MultiMapping==4.1
multipart==0.2.4
PasteDeploy==3.0.1
Persistence==3.6
persistent==4.9.3
protobuf==4.21.12
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
python-gettext==4.1
pytz==2022.6
RestrictedPython==6.0
roman==3.3
six==1.16.0
soupsieve==2.3.2.post1
transaction==3.0.1
typing_extensions==4.4.0
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
WSGIProxy2==0.5.1
z3c.pt==3.3.1
zc.lockfile==2.0
ZConfig==3.6.1
zExceptions==4.2
ZODB==5.8.0
zodbpickle==2.6
Zope==5.7
zope.annotation==4.8
zope.browser==2.4
zope.browsermenu==4.4
zope.browserpage==4.4.0
zope.browserresource==4.4
zope.cachedescriptors==4.4
zope.component==5.0.1
zope.configuration==4.4.1
zope.container==4.10
zope.contentprovider==4.2.1
zope.contenttype==4.6
zope.datetime==4.3.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.dottedname==5.0
zope.event==4.6
zope.exceptions==4.6
zope.filerepresentation==5.0.0
zope.globalrequest==1.6
zope.hookable==5.4
zope.i18n==4.9.0
zope.i18nmessageid==5.1.1
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.location==4.3
zope.pagetemplate==4.6.0
zope.processlifetime==2.4
zope.proxy==4.6.1
zope.ptresource==4.3.0
zope.publisher==6.1.0
zope.schema==6.2.1
zope.security==5.8
zope.sequencesort==4.2
zope.site==4.6.1
zope.size==4.4
zope.structuredtext==4.4
zope.tal==4.5
zope.tales==5.2
zope.testbrowser==5.6.1
zope.testing==4.10
zope.traversing==4.4.1
zope.viewlet==4.3
